+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - プロダクトオファー承認時のアクセスコード指定を4桁の数字で固定
 - 所有権検索条件拡張
 - ペイメントカードに対する所有権検索時に、ペイメントカードタイプを動的に検索するように調整
+- sskts専用メンバーシップ登録時に、agent.additionalPropertyを指定できるように調整
 
 ### Deprecated